introduce vm/modules/ and move builtins there
