+++ spy/backend/c/wrapper.py
@@ -10,7 +10,7 @@ from spy.vm.str import ll_spy_Str_new
 from spy.vm.module import W_Module
 from spy.vm.function import W_Func, W_FuncType
 from spy.vm.vm import SPyVM
-from spy.vm.builtins import B
+from spy.vm.b import B
 
 
 class WasmModuleWrapper:
