+++ spy/tests/compiler/test_00_bluemod.py
@@ -2,7 +2,7 @@ from typing import Any, Literal, Optional
 import textwrap
 import pytest
 from spy.vm.vm import SPyVM
-from spy.vm.builtins import B
+from spy.vm.b import B
 from spy.vm.function import W_FuncType, W_ASTFunc
 from spy.backend.interp import InterpModuleWrapper
