+++ spy/vm/astframe.py
@@ -7,7 +7,7 @@ from spy.location import Loc
 from spy.errors import (SPyRuntimeAbort, SPyTypeError, SPyNameError,
                         SPyRuntimeError, maybe_plural)
 from spy.irgen.symtable import Symbol
-from spy.vm.builtins import B
+from spy.vm.b import B
 from spy.vm.object import W_Object, W_Type, W_i32, W_bool
 from spy.vm.str import W_str
 from spy.vm.function import W_Func, W_FuncType, W_ASTFunc, Namespace
