+++ spy/vm/function.py
@@ -58,7 +58,7 @@ class W_FuncType(W_Type):
         It's meant to be used in tests, it's not robust at all, especially in
         case of wrong inputs. It supports only builtin types.
         """
-        from spy.vm.builtins import B
+        from spy.vm.b import B
         def parse_type(s: str) -> Any:
             return getattr(B, f'w_{s}')
