+++ spy/vm/ops.py
@@ -1,5 +1,5 @@
 from typing import TYPE_CHECKING, Any, Optional
-from spy.vm.builtins import B
+from spy.vm.b import B
 from spy.vm.str import W_str
 from spy.vm.object import W_Object, W_Type, W_i32, W_bool
 from spy.vm.function import W_FuncType, W_Func
